requirements/development.txt
asgiref==3.8.1
asttokens==2.2.1
backcall==0.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
debugpy==1.6.4
defusedxml==0.8.0rc2
Django==4.2.15
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
entrypoints==0.4
exceptiongroup==1.2.2
executing==1.2.0
idna==3.7
iniconfig==2.0.0
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.4
pluggy==1.5.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.22
Pygments==2.13.0
PyJWT==2.9.0
pyparsing==3.0.9
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pywin32==305
pyzmq==24.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.6.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.5
